fix: split Popover components and fixed issue
