+++ packages/react-styled-ui/src/Popover/PopoverFooter.js
@@ -1,0 +1,9 @@
+import Box from '../Box';
+
+const PopoverFooter = props => (
+  <Box as="footer" pt="4x" {...props} />
+);
+
+PopoverFooter.displayName = 'PopoverFooter';
+
+export default PopoverFooter;
