+++ packages/react-styled-ui/src/Popover/PopoverHeader.js
@@ -1,0 +1,21 @@
+import { usePopover } from './context';
+import PseudoBox from '../PseudoBox';
+import { usePopoverHeaderStyle } from './styles';
+
+const PopoverHeader = props => {
+  const { headerId } = usePopover();
+  const headerStyleProps = usePopoverHeaderStyle();
+
+  return (
+    <PseudoBox
+      as="header"
+      id={headerId}
+      {...headerStyleProps}
+      {...props}
+    />
+  );
+};
+
+PopoverHeader.displayName = 'PopoverHeader';
+
+export default PopoverHeader;
